refactor: refactor quote handling by moving QuoteType class to its own file
